requirements.txt
langchain>=0.1.0
langgraph>=0.0.55
langchain-openai>=0.0.8
langchain-community>=0.0.38
langchain-experimental>=0.0.45
sqlalchemy>=2.0.25
pandas>=2.1.4
numpy>=1.24.0
openpyxl>=3.1.2
lxml>=4.9.3
tabulate>=0.9.0
pywin32>=308; sys_platform == "win32"
exchangelib>=5.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
pymysql>=1.1.0